requirements.txt
astroid==2.3.1
Click==7.0
coverage==4.5.4
Flask==1.1.1
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.2
pymongo==3.9.0
python-dotenv==0.10.3
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
